+++ src/providers/__init__.py
@@ -1,7 +1,8 @@
 #!/usr/bin/python3
 
 from src.providers.ollama import OllamaProvider
+from src.providers.openai_compatible import OpenAICompatibleProvider
 from src.providers.openrouter import OpenRouterProvider
 from src.providers.utils import get_requirements
 
-__all__ = ["OllamaProvider", "OpenRouterProvider", "get_requirements"]
+__all__ = ["OllamaProvider", "OpenAICompatibleProvider", "OpenRouterProvider", "get_requirements"]